requirements.txt
pandas
numpy
PyPDF2
python-docx
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pymongo
python-dotenv
langchain
langchain-google-genai
flask